+++ src/containers/either/index.ts
@@ -1,5 +1,3 @@
 import Either from './either';
-import Left from './left';
-import Right from './right';
 
 export default Either;