+++ src/types.ts
@@ -1,3 +1,5 @@
 export type Map<I, O> = (value: I) => O;
 
-export type Object = { [key: string]: any };
+export type O = { [key: string]: unknown };
+
+export type F = (...args: unknown[]) => unknown;