+++ src/utils/inspect/inspect.ts
@@ -1,6 +1,8 @@
-import { Object } from 'types';
+import { O, F } from 'types';
 
-const inspectFunction = (fn: Function): string => (
+import { ValueWithInspect } from './types';
+
+const inspectFunction = (fn: F): string => (
   `function ${fn.name}() { }` || fn.toString()
 );
 
@@ -8,39 +10,39 @@ const inspectString = (term: string): string => (
   `'${term}'`
 );
 
-const inspectObject = (value: Object): string => {
+const inspectObject = (value: O): string => {
   const subTerms = Object.keys(value).map((key) => [key, inspect(value[key])]);
   const result = subTerms.map((subTerm) => subTerm.join(': ')).join(', ');
 
   return `{ ${result} }`;
 };
 
-const inspectArray = (value: any[]): string => {
+const inspectArray = (value: unknown[]): string => {
   const result = value.map(inspect).join(', ');
 
   return `[ ${result} ]`;
 };
 
-function inspect(value: any): string {
+function inspect(value: unknown): string {
   const isArray = Array.isArray(value);
   const isString = typeof value === 'string';
   const isObject = typeof value === 'object';
   const isFunction = typeof value === 'function';
-  const hasInspectMethod = typeof value?.inspect === 'function';
+  const hasInspectMethod = typeof (<O>value)?.inspect === 'function';
 
   if (hasInspectMethod) {
-    return value.inspect();
+    return (<ValueWithInspect>value).inspect();
   } else if (isFunction) {
-    return inspectFunction(value);
+    return inspectFunction(<F>value);
   } else if (isArray) {
-    return inspectArray(value);
+    return inspectArray(<unknown[]>value);
   } else if (isString) {
-    return inspectString(value);
+    return inspectString(<string>value);
   } else if (isObject) {
-    return inspectObject(value);
+    return inspectObject(<O>value);
   } else {
     return String(value);
   }
-};
+}
 
 export default inspect;