+++ src/utils/is-nil/is-nil.ts
@@ -1,4 +1,4 @@
-const isNil = (value: any): boolean => {
+const isNil = (value: unknown): boolean => {
   const isNull = value === null;
   const isUndefined = value === undefined;