+++ src/utils/prop/prop.ts
@@ -1,8 +1,8 @@
 import { curry } from 'utils';
-import { Object } from 'types';
+import { O } from 'types';
 
-const prop = <T>(key: string, obj: Object): T => (
-  obj[key]
+const prop = <T>(key: string, obj: O): T => (
+  <T>obj[key]
 );
 
 export default curry(prop);