+++ CHANGELOG.md
@@ -4,3 +4,15 @@
 
 ## 0.0.1 - 2021-02-24
 - Initial release
+- Added `Container` container
+- Added `Either` container
+- Added `Maybe` container
+- Added `compose` util
+- Added `curry` util
+- Added `either` util
+- Added `inspect` util
+- Added `isNil` util
+- Added `left` util
+- Added `prop` util
+- Added `Container` interface
+- Added `Either` interface